sql: dissallow bindings for DDL
Bindings could not be used in stored ASTs because they allocate
memory registers and process assignments during parsing(temporal register
should be allocated for each variable in parse context, but it is temporal).
Original sqlite3 did validations that AST to be persisted doesn't have
auto-assigment Varibles on triggers and checks creation.
On DDL integration completion we've got rid this mechanism.
Now it should be returned.
We use flag 'parse_only' which is set by parser on compiling AST to
determine attempt to use bindings in DDL(triggers, defaults and checks
creation) that is incorrect and raise an error.

Closes #3653.