promote: expose initial promotion to internal Lua API
Now for a user it is a problem to turn its existing cluster
into managed by promotion protocol. Box.ctl.promote() just does
not work if no old master is found. So this patch exposes a
function box.internal.initial_promote() to a user land to help
in first promotion. This is a temporary API until a promotion
is able to withstand master absence.

Besides, this API can be used for a while for a regular failover.
If an old master is dead, a user can reset promotion history
by box.ctl.promote_reset() and do initial promotion on another
node. But now there is no way to kill an old master if it
suddenly resurrects.

Follow up #3055